build-profile.py: Initial commit.
This is a Python script that will generate a mobile profile suitable for
use with Apple devices running iOS 9 or newer.

It defaults to using AES-256 for encryption, SHA2-256 for integrity,
and 2048-bit Diffie Hellman group 14 for both IKE and Child SA parameters.

Any of the above are overridable on the command line.